restructure core into hexagonal architecture
