+++ dcs/adapters/outbound/s3.py
@@ -16,4 +16,4 @@
 """Implementation of object storage adapters."""
 
 # pylint: disable=unused-import
-from hexkit.providers.s3 import S3ObjectStorage, S3Config  # noqa: F401
+from hexkit.providers.s3 import S3Config, S3ObjectStorage  # noqa: F401
