+++ tests/fixtures/__init__.py
@@ -16,6 +16,5 @@
 """Fixtures that can be used in both unit and integration tests"""
 
 from .config import DEFAULT_CONFIG, get_config  # noqa: F401
-from .psql import psql_fixture  # noqa: F401
 from .s3 import s3_fixture  # noqa: F401
 from .state import FILES  # noqa: F401
